chore: renamed radiogroup from radio-group
