+++ packages/components/src/components/radiogroup/__screenshots__/chrome/mdc-radio-group-high-contrast.png
