+++ packages/components/src/components/radiogroup/__screenshots__/chrome/mdc-radio-group-ltr.png
