+++ packages/components/src/components/radiogroup/__screenshots__/firefox/mdc-radio-group-ltr.png
