+++ packages/components/src/components/radiogroup/__screenshots__/firefox/mdc-radio-group-rtl.png
