+++ packages/components/src/components/radiogroup/__screenshots__/mobile-chrome/mdc-radio-group-rtl.png
