+++ packages/components/src/components/radiogroup/__screenshots__/mobile-safari/mdc-radio-group-rtl.png
