+++ packages/components/src/components/radiogroup/__screenshots__/msedge/mdc-radio-group-high-contrast.png
