+++ packages/components/src/components/radiogroup/__screenshots__/msedge/mdc-radio-group-ltr.png
