+++ packages/components/src/components/radiogroup/__screenshots__/tablet-chrome/mdc-radio-group-high-contrast.png
