+++ packages/components/src/components/radiogroup/__screenshots__/tablet-chrome/mdc-radio-group-ltr.png
