+++ packages/components/src/components/radiogroup/__screenshots__/tablet-safari/mdc-radio-group-ltr.png
