+++ packages/components/src/components/radiogroup/__screenshots__/webkit/mdc-radio-group-ltr.png
