+++ packages/components/src/index.ts
@@ -20,7 +20,7 @@ import Tab from './components/tab';
 import FormfieldGroup from './components/formfieldgroup';
 import { inMemoryCache, webAPIIconsCache } from './utils/icon-cache';
 import Spinner from './components/spinner';
-import RadioGroup from './components/radio-group';
+import RadioGroup from './components/radiogroup';
 
 import type { TextType } from './components/text/text.types';
 import type { SpinnerSize, SpinnerVariant } from './components/spinner/spinner.types';
